refactor: extractDefaultFieldsFromFilters to filters utils
