+++ apps/nextjs-app/src/features/app/blocks/view/grid/GridViewBase.tsx
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import type { IFieldVo, IFilterItem, IFilterSet } from '@teable/core';
+import type { IFieldVo } from '@teable/core';
 import {
   FieldKeyType,
   RowHeightLevel,
@@ -46,6 +46,7 @@ import {
   useGridPrefillingRow,
   SelectableType,
 } from '@teable/sdk';
+import { extractDefaultFieldsFromFilters } from '@teable/sdk/components/filter/utils';
 import { GRID_DEFAULT } from '@teable/sdk/components/grid/configs';
 import { useScrollFrameRate } from '@teable/sdk/components/grid/hooks';
 import {
@@ -84,33 +85,6 @@ interface IGridViewProps {
 }
 
 const { scrollBuffer, columnAppendBtnWidth } = GRID_DEFAULT;
-
-function extractDefaultFieldsFromFilters(filters: IFilterSet | null | undefined): {
-  [fieldId: string]: unknown;
-} {
-  const result: { [fieldId: string]: unknown } = {};
-  const repeatedFieldIds = new Set<string>();
-
-  function handleFilterItem(filter: IFilterItem | null | undefined) {
-    if (filter?.operator === 'is' && filter.fieldId) {
-      if (filter.fieldId in result) {
-        // mark as repeat and delete
-        delete result[filter.fieldId];
-        repeatedFieldIds.add(filter.fieldId);
-      } else if (!repeatedFieldIds.has(filter.fieldId)) result[filter.fieldId] = filter.value;
-    }
-  }
-
-  // recursively traverse the filters object
-  function traverse(filter: IFilterSet | IFilterItem | null | undefined) {
-    if (filter && 'filterSet' in filter) filter.filterSet.forEach(traverse);
-    else if (filter) handleFilterItem(filter);
-  }
-
-  traverse(filters);
-
-  return result;
-}
 
 export const GridViewBase: React.FC<IGridViewProps> = (props: IGridViewProps) => {
   const { onRowExpand } = props;
